moving preprocessing and augmentation under data for consistency
